+++ lib-SLExternal/assimp/code/IFCOpenings.cpp
@@ -365,7 +365,7 @@ void InsertWindowContours(const ContourVector& contours,
 					}
 
 					if (const size_t d = curmesh.verts.size()-old) {
-						curmesh.vertcnt.push_back(d);
+						curmesh.vertcnt.push_back((unsigned int)d);
 						std::reverse(curmesh.verts.rbegin(),curmesh.verts.rbegin()+d);
 					}
 					if (n == very_first_hit) {
@@ -550,7 +550,7 @@ void CleanupOuterContour(const std::vector<IfcVector2>& contour_flat, TempMesh& 
 				clipper.Execute(ClipperLib::ctIntersection,clipped,ClipperLib::pftNonZero,ClipperLib::pftNonZero);
 
 				BOOST_FOREACH(const ClipperLib::ExPolygon& ex, clipped) {
-					iold.push_back(ex.outer.size());
+					iold.push_back((unsigned int)ex.outer.size());
 					BOOST_FOREACH(const ClipperLib::IntPoint& point, ex.outer) {
 						vold.push_back(IfcVector3(
 							from_int64(point.X),
