+++ lib-SLExternal/assimp/code/Subdivision.cpp
@@ -192,7 +192,7 @@ void CatmullClarkSubdivider::Subdivide (
 		}
 
 		outmeshes.push_back(NULL);inmeshes.push_back(i);
-		maptbl.push_back(s);
+		maptbl.push_back((unsigned int)s);
 	}
 
 	// Do the actual subdivision on the preallocated storage. InternSubdivide
